Add support for Jersey's JResponse as an alternative to @TypeHint
When a custom Response is used in a JAX-RS service, enunciate is not able to introspect the returned type unless the TypeHint annotation is used. This might not be acceptable for a project that cannot take a compile and runtime dependency on the enunciate annotations library.

This change provides an alternative by adding introspection of Jersey's JResponse. JResponse is an alternative to javax.ws.rs.core.Response that allows the entity type to be specified as a generic type argument.

Example:

@POST
public JResponse<MyObject> createMyObject(...)

is equivalent to

@POST
@TypeHint(MyObject.class)
public JResponse createMyObject(...)