Add format information to Swagger data types
Swagger output can be used to generate source code, which is really
handy for use with JAX-RS 2 clients.

For this to work properly, the full type of the simple integer and and
floatingpoint types need to be correctly preserved. Otherwise the
generated code will be using the wrong (simplified) types.

This code uses the type design in the swagger specification:
http://swagger.io/specification/
It might be worth to add extra types, when/if required by other
backends.