Add methods used by swagger/openapi backend to Property interface
Swagger backend references methods in the jaxb module which are
not exposed by the Java API.
I guess it works because freemarker uses reflection on the provided
beans to make calls.

In the OpenAPI backend I have moved some logic from the freemarker
template into Java classes. So I need the calls to be exposed via
the proper interfaces.