Add support for nested arrays
If a domain class has a field of type multi dimensional array, e.g.
double[][], then in the documentation the type was incorrect (in
this example it ends up as double[] in the HTML).

This commit makes sure that in this case the HTML contains the correct
type, double[][]. This only works when the getter of the field contains
the @TypeHint annotation, e.g. @TypeHint(double[][].class.